requirements.txt
slackclient>=2.0.0
slackeventsapi>=2.1.0
